AgencyOS/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
django-extensions==3.2.3
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
coverage==7.6.1
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1

# ConnectOS shared platform layer
connectos-platform @ file:../../ConnectOS/platform
